requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_pytorch_select=0.1=cpu_0
_tflow_select=2.3.0=eigen
absl-py=0.11.0=py38haa244fe_0
aiohttp=3.7.3=py38h294d835_1
anyio=2.0.2=py38haa244fe_4
argon2-cffi=20.1.0=py38he774522_1
astor=0.8.1=pyh9f0ad1d_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=3.0.1=py_1000
async_generator=1.10=pyhd3eb1b0_0
attrs=20.3.0=pyhd3eb1b0_0
babel=2.9.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bamnostic=1.1.6=pyh44b312d_1
beautifulsoup4=4.9.3=pypi_0
bert-tensorflow=1.0.4=pypi_0
biopython=1.78=py38h294d835_2
blas=1.0=mkl
bleach=3.2.3=pyhd3eb1b0_0
blinker=1.4=py_1
boto3=1.17.52=pypi_0
botocore=1.20.52=pypi_0
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2020.12.5=h5b45459_0
cachetools=4.2.1=pyhd8ed1ab_0
certifi=2020.12.5=py38haa244fe_1
cffi=1.14.4=py38hcd4344a_0
chardet=3.0.4=py38h9bdc248_1008
click=7.1.2=pyh9f0ad1d_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=3.3.1=py38hcd4344a_0
cudatoolkit=10.0.130=0
cycler=0.10.0=py38_0
cython=0.29.21=py38ha925a31_0
dataclasses=0.8=pyh6d0b6a4_7
decorator=4.4.2=pyhd3eb1b0_0
defusedxml=0.6.0=py_0
dill=0.3.3=pyhd3eb1b0_0
entrypoints=0.3=py38_0
filelock=3.0.12=pyhd3eb1b0_1
flatbuffers=1.12=pypi_0
freetype=2.10.4=hd328e21_0
gast=0.3.3=pypi_0
google-auth=1.24.0=pyhd3deb0d_0
google-auth-oauthlib=0.4.2=pypi_0
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.32.0=pypi_0
h5py=2.10.0=nompi_py38h6053941_105
hdf5=1.10.6=nompi_h5268f04_1114
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.19.0=py38hd4e2768_1
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py38_0
jinja2=2.11.2=pyhd3eb1b0_0
jmespath=0.10.0=pypi_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
json5=0.9.5=py_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.0=py38haa95532_0
jupyter_server=1.2.2=py38haa244fe_1
jupyterlab=3.0.6=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.1.3=pyhd8ed1ab_0
keras=2.4.3=0
keras-applications=1.0.8=py_1
keras-base=2.4.3=py_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.1=py38hd77b12b_0
krb5=1.17.2=hbae68bd_0
libcurl=7.71.1=h4b64cdc_8
libgpuarray=0.7.6=hfa6e2cd_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.14.0=h7755175_0
libpython=2.1=py38_0
libsodium=1.0.18=h62dcd97_0
libssh2=1.9.0=hb06d900_5
libtiff=4.1.0=h56a325e_1
libuv=1.40.0=he774522_0
lz4-c=1.9.3=h2bbff1b_0
m2w64-binutils=2.25.1=5
m2w64-bzip2=1.0.6=6
m2w64-crt-git=5.0.0.4636.2595836=2
m2w64-gcc=5.3.0=6
m2w64-gcc-ada=5.3.0=6
m2w64-gcc-fortran=5.3.0=6
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gcc-objc=5.3.0=6
m2w64-gmp=6.1.0=2
m2w64-headers-git=5.0.0.4636.c0ad18a=2
m2w64-isl=0.16.1=2
m2w64-libiconv=1.14=6
m2w64-libmangle-git=5.0.0.4509.2e5a9a2=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
m2w64-make=4.1.2351.a80a8b8=2
m2w64-mpc=1.0.3=3
m2w64-mpfr=3.1.4=4
m2w64-pkg-config=0.29.1=2
m2w64-toolchain=5.3.0=7
m2w64-tools-git=5.0.0.4592.90b8472=2
m2w64-windows-default-manifest=6.4=3
m2w64-winpthreads-git=5.0.0.4634.697f757=2
m2w64-zlib=1.2.8=10
mako=1.1.4=pyhd3eb1b0_0
markdown=3.3.3=pyh9f0ad1d_0
markupsafe=1.1.1=py38he774522_0
matplotlib=3.3.4=py38haa95532_0
matplotlib-base=3.3.4=py38h49ac443_0
mistune=0.8.4=py38he774522_1000
mkl=2020.2=256
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.2.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
msys2-conda-epoch=20160418=1
multidict=5.1.0=py38h294d835_1
nbclassic=0.2.6=pyhd8ed1ab_0
nbclient=0.5.1=py_0
nbconvert=6.0.7=py38_0
nbformat=5.1.2=pyhd3eb1b0_1
nest-asyncio=1.4.3=pyhd3eb1b0_0
ninja=1.10.2=py38h6d14046_0
notebook=6.2.0=py38haa95532_0
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
openssl=1.1.1k=h8ffe710_0
opt-einsum=3.3.0=pypi_0
opt_einsum=3.3.0=py_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.4=py38hd77b12b_0
pandoc=2.11=h9490d1a_0
pandocfilters=1.4.3=py38haa95532_1
parso=0.8.1=pyhd3eb1b0_0
patsy=0.5.1=py_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.0=py38h4fa10fc_0
pip=20.3.3=py38haa95532_0
prometheus_client=0.9.0=pyhd3eb1b0_0
prompt-toolkit=3.0.8=py_0
protobuf=3.14.0=pypi_0
pyahocorasick=1.4.0=py38h1e8a9f7_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pygments=2.7.4=pyhd3eb1b0_0
pygpu=0.7.6=py38h2a96729_0
pyjwt=2.0.1=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38ha925a31_4
pyreadline=2.1=py38haa244fe_1003
pyrsistent=0.17.3=py38he774522_0
pysocks=1.7.1=py38haa95532_0
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.8=1_cp38
pytorch=1.4.0=cpu_py38ha775e86_0
pytz=2020.5=pyhd3eb1b0_0
pywin32=227=py38he774522_1
pywinpty=0.5.7=py38_0
pyyaml=5.4.1=py38h2bbff1b_1
pyzmq=20.0.0=py38hd77b12b_1
qt=5.9.7=vc14h73c81de_0
regex=2021.4.4=py38h2bbff1b_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=pyh9f0ad1d_0
rsa=4.7=pyhd3deb0d_0
s3transfer=0.3.7=pypi_0
sacremoses=0.0.43=pypi_0
scipy=1.5.2=py38h14eb087_0
seaborn=0.11.1=pyhd3eb1b0_0
send2trash=1.5.0=pyhd3eb1b0_1
sentencepiece=0.1.91=pypi_0
setuptools=52.0.0=py38haa95532_0
sip=4.19.13=py38ha925a31_0
six=1.15.0=py38haa95532_0
sniffio=1.1.0=py38_2
soupsieve=2.1=pypi_0
sqlite=3.33.0=h2a8f88b_0
statsmodels=0.12.2=py38h347fdf6_0
tensorboard=2.4.1=pyhd8ed1ab_0
tensorboard-plugin-wit=1.8.0=pyh44b312d_0
tensorflow=2.4.1=pypi_0
tensorflow-base=2.3.0=eigen_py38h75a453f_0
tensorflow-estimator=2.4.0=pyh9656e83_0
tensorflow-hub=0.12.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.2=py38haa95532_0
testpath=0.4.4=pyhd3eb1b0_0
theano=1.0.4=py38_0
tk=8.6.10=he774522_0
tokenizers=0.5.0=pypi_0
torchvision=0.2.2=py_3
tornado=6.1=py38h2bbff1b_0
tqdm=4.59.0=pyhd3eb1b0_1
traitlets=5.0.5=pyhd3eb1b0_0
transformers=2.5.0=dev_0
tweepy=3.10.0=pypi_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.3=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
weblogo=3.7.8=pypi_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38_0
winpty=0.4.3=4
wrapt=1.12.1=py38h294d835_3
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yarl=1.6.3=py38h294d835_1
zeromq=4.3.3=ha925a31_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
